refactor(frontend): decompose useCodeIntelAnalysis.ts into sub-composables (#2260)
Split 1,923-line monolith into 6 focused files while preserving the
same public API surface via a facade pattern:

- codeIntelTypes.ts: shared type definitions (372 lines)
- useCodeIntelScores.ts: security/performance/redis scores (323 lines)
- useCodeSmellAnalysis.ts: code smells + health score (173 lines)
- useBugPrediction.ts: bug prediction + risk analysis (230 lines)
- useSpecializedAnalysis.ts: API endpoints, config duplicates,
  environment, ownership, cross-language analysis (583 lines)
- useCodeIntelAnalysis.ts: facade importing sub-composables (233 lines)
